added other controllers

+++ Inventory.DataAccess/Models/OrderDetail.cs
@@ -9,10 +9,10 @@ namespace Inventory.DataAccess.Models
         public int OrderDetailId { get; set; }
         public int OrderId { get; set; }
         public int ItemId { get; set; }
-        public decimal Qty { get; set; }
-        public decimal Discount { get; set; }
-        public decimal Tax { get; set; }
-        public decimal Total { get; set; }
+        public double Qty { get; set; }
+        public double Discount { get; set; }
+        public double Tax { get; set; }
+        public double Total { get; set; }
         public string Remarks { get; set; }
         public int CompanyId { get; set; }
         public bool IsVoid { get; set; }
